Refactored tests into Payload, Metric and Template
